Revert "Merge branch 'master' of https://github.com/Caephler/GooreatRestaurant"
This reverts commit f8efb938f81b27470efbba5a243bd4ea4dbcce44.

+++ slime/Models/Stage/Order.swift
@@ -9,10 +9,10 @@
 import UIKit
 
 class Order {
-    var timeLimit: CGFloat
+    var timeLimit: Int
     let recipeWanted: Recipe
 
-    init(_ recipe: Recipe, withinTime time: CGFloat) {
+    init(_ recipe: Recipe, withinTime time: Int) {
         self.timeLimit = time
         self.recipeWanted = recipe
     }
